+++ blog/views.py
@@ -1,39 +1,7 @@
 from django.http import  HttpRequest, Http404
 from django.shortcuts import render
-from rest_framework.generics import (ListAPIView,
-                                    RetrieveAPIView,
-                                    CreateAPIView,
-                                    UpdateAPIView,
-                                    DestroyAPIView)
-
-from .serializers import ArticleSerializer
 
 from blog.models import Article, Category
-
-
-class ArticleListAPIView(ListAPIView):
-    queryset = Article.objects.all()
-    serializer_class = ArticleSerializer
-
-
-class ArticleDetailAPIView(RetrieveAPIView):
-    queryset = Article.objects.all()
-    serializer_class = ArticleSerializer
-
-
-class ArticleCreateAPIView(CreateAPIView):
-    queryset = Article.objects.all()
-    serializer_class = ArticleSerializer
-
-
-class ArticleUpdateAPIView(UpdateAPIView):
-    queryset = Article.objects.all()
-    serializer_class = ArticleSerializer
-
-
-class ArticleDeleteAPIView(DestroyAPIView):
-    queryset = Article.objects.all()
-    serializer_class = ArticleSerializer
 
 
 def home_page(request: HttpRequest):
