+++ dj_hw/urls.py
@@ -21,4 +21,5 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('blog.urls')),
     path('', include('user.urls')),
+    path('', include('api.urls')),
 ]
